+++ contracts/factory/src/tests/queries.rs
@@ -1,5 +1,5 @@
 use super::setup::{deploy_factory_contract, generate_lp_init_info};
-use crate::storage::{Asset, LpPortfolio, Stake, StakePortfolio, UserPortfolio};
+use crate::storage::{Asset, LpPortfolio, PoolType, Stake, StakePortfolio, UserPortfolio};
 use crate::tests::setup::{lp_contract, stake_contract};
 use crate::token_contract;
 use phoenix::utils::{LiquidityPoolInitInfo, StakeInitInfo, TokenInitInfo};
@@ -136,18 +136,24 @@ fn test_deploy_multiple_liquidity_pools() {
         &first_lp_init_info,
         &String::from_str(&env, "Pool"),
         &String::from_str(&env, "PHO/BTC"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
     let second_lp_contract_addr = factory.create_liquidity_pool(
         &admin.clone(),
         &second_lp_init_info,
         &String::from_str(&env, "Pool #2"),
         &String::from_str(&env, "PHO/ETH"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
     let third_lp_contract_addr = factory.create_liquidity_pool(
         &admin.clone(),
         &third_lp_init_info,
         &String::from_str(&env, "Pool #3"),
         &String::from_str(&env, "PHO/XLM"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let first_result = factory.query_pool_details(&lp_contract_addr);
@@ -336,18 +342,24 @@ fn test_queries_by_tuple() {
         &first_lp_init_info,
         &String::from_str(&env, "Pool"),
         &String::from_str(&env, "PHO/BTC"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
     let second_lp_contract_addr = factory.create_liquidity_pool(
         &admin.clone(),
         &second_lp_init_info,
         &String::from_str(&env, "Pool #2"),
         &String::from_str(&env, "PHO/ETH"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
     let third_lp_contract_addr = factory.create_liquidity_pool(
         &admin.clone(),
         &third_lp_init_info,
         &String::from_str(&env, "Pool #3"),
         &String::from_str(&env, "PHO/XLM"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let first_result = factory.query_pool_details(&lp_contract_addr);
@@ -438,10 +450,10 @@ fn test_query_user_portfolio_with_stake() {
     let factory = deploy_factory_contract(&env, Some(admin.clone()));
 
     let first_lp_init_info = generate_lp_init_info(
-        &token1,
-        &token2,
+        token1.address.clone(),
+        token2.address.clone(),
         manager.clone(),
-        &admin,
+        admin.clone(),
         fee_recipient.clone(),
     );
 
@@ -450,6 +462,8 @@ fn test_query_user_portfolio_with_stake() {
         &first_lp_init_info,
         &String::from_str(&env, "Pool"),
         &String::from_str(&env, "PHO/BTC"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let first_lp_client = lp_contract::Client::new(&env, &first_lp_contract_addr);
@@ -532,14 +546,21 @@ fn test_query_user_portfolio_with_stake() {
         }
     );
 
-    let second_lp_init_info =
-        generate_lp_init_info(&token3, &token4, manager.clone(), &admin, fee_recipient);
+    let second_lp_init_info = generate_lp_init_info(
+        token3.address.clone(),
+        token4.address.clone(),
+        manager.clone(),
+        admin.clone(),
+        fee_recipient,
+    );
 
     let second_lp_contract_addr = factory.create_liquidity_pool(
         &admin.clone(),
         &second_lp_init_info,
         &String::from_str(&env, "Second Pool"),
         &String::from_str(&env, "PHO/ETH"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let second_lp_client = lp_contract::Client::new(&env, &second_lp_contract_addr);
@@ -663,10 +684,10 @@ fn test_query_user_portfolio_with_multiple_users_staking_in_multiple_liquidity_p
 
     // first liquidity pool
     let first_lp_init_info = generate_lp_init_info(
-        &token1,
-        &token2,
+        token1.address.clone(),
+        token2.address.clone(),
         manager.clone(),
-        &admin,
+        admin.clone(),
         fee_recipient.clone(),
     );
 
@@ -675,6 +696,8 @@ fn test_query_user_portfolio_with_multiple_users_staking_in_multiple_liquidity_p
         &first_lp_init_info,
         &String::from_str(&env, "Pool"),
         &String::from_str(&env, "PHO/BTC"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let first_lp_client = lp_contract::Client::new(&env, &first_lp_contract_addr);
@@ -688,10 +711,10 @@ fn test_query_user_portfolio_with_multiple_users_staking_in_multiple_liquidity_p
 
     // second liquidity pool
     let second_lp_init_info = generate_lp_init_info(
-        &token3,
-        &token4,
+        token3.address.clone(),
+        token4.address.clone(),
         manager.clone(),
-        &admin,
+        admin.clone(),
         fee_recipient.clone(),
     );
 
@@ -700,6 +723,8 @@ fn test_query_user_portfolio_with_multiple_users_staking_in_multiple_liquidity_p
         &second_lp_init_info,
         &String::from_str(&env, "Second Pool"),
         &String::from_str(&env, "PHO/ETH"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let second_lp_client = lp_contract::Client::new(&env, &second_lp_contract_addr);
@@ -1183,6 +1208,8 @@ fn test_query_user_portfolio_without_stake() {
         &first_lp_init_info,
         &String::from_str(&env, "Pool"),
         &String::from_str(&env, "PHO/BTC"),
+        &PoolType::Xyk,
+        &None::<u64>,
     );
 
     let first_lp_client = lp_contract::Client::new(&env, &lp_contract_addr);